finishes base version with regular type of pipeline
