+++ profiling/README.md
@@ -1,0 +1,3 @@
+# Profiling Event Pipeline
+
+This section provides a list of tools that help profile the performance of the various types of pipelines that can be created using event-pipeline. It is built to be highly configurable such that running the different configurations can be done easily using command line flags.
